+++ tests/Floats/FloatsTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace OceanMoon\Core\Tests\Floats;
 
 use DomainException;
-use OceanMoon\Core\ExponentFormat;
-use OceanMoon\Core\FloatFormat;
+use OceanMoon\Core\Enums\ExponentFormat;
+use OceanMoon\Core\Enums\FloatFormat;
 use OceanMoon\Core\Floats;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;